% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dHyperTPoisson}
\alias{dHyperTPoisson}
\title{Probability function for Hypergeometric Truncated Poisson distribution}
\usage{
dHyperTPoisson(nhat, lambda, alpha, N, log = FALSE)
}
\arguments{
\item{nhat}{The observed number.}

\item{lambda}{The rate parameter for the poisson distribution.}

\item{alpha}{The sampling percentage.}

\item{N}{The population size.}

\item{log}{Should the log probability be returned.}
}
\description{
Probability function for Hypergeometric Truncated Poisson distribution
}
\examples{
dHyperTPoisson(nhat=2, lambda=2, alpha=1, N=10)
dHyperTPoisson(1, lambda=c(2,3), alpha=1, N=10)
dHyperTPoisson(1:3, lambda=2, alpha=1, N=10)
dHyperTPoisson(1:3, lambda=2, alpha=1, N=10, log=TRUE)

}
