% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distributions.R
\name{dtpois}
\alias{dtpois}
\alias{ptpois}
\alias{qtpois}
\alias{rtpois}
\title{Zero Truncated Poisson Distribution}
\usage{
dtpois(x, lambda)

ptpois(q, lambda)

qtpois(p, lambda)

rtpois(n, lambda)
}
\description{
A poisson distribution with the zero removed.
}
\examples{
dtpois(2, lambda=4)
dtpois(0, lambda=4)
dtpois(1:100, lambda=4) \%>\% sum()
}
