% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePopulation.R
\name{generate_population}
\alias{generate_population}
\title{Create a population of HAI cases}
\usage{
generate_population(n, lambda, mix.frac = 0.01)
}
\arguments{
\item{n}{- Total number of patients that we could have sampled}

\item{lambda}{- Poisson paramter}

\item{mix.frac}{Mixing fraction of LogNoramal to be included}
}
\value{
A data frame of approximately n patients, each with a ClusterID
        and if their infection was Community or Hospital aquired
}
\description{
Create a population of HAI cases
}
