% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prop2Perc.R
\name{prop2perc}
\alias{prop2perc}
\title{Convert proportions to percent labels}
\usage{
prop2perc(input)
}
\arguments{
\item{input}{A vector of numeric values representing proportions [0,1].}
}
\value{
A vector of factors values as percentages, arranged in assending order.
}
\description{
Convert proportions to percent labels
}
